fix(vi-mode): fix `^` motion
Previously the `^` motion was behaving like the `0` motion which is to
move the cursor to the start of the line. But in vim, the `^` motion is
different from `0` motion. In vim `^` moved the cursor to the first
non-blank character of the line. (See `:h ^`)

Updated test case for `d^` and `c^` command since it has different
behaviour now.